Nest xenophile's TypeScript model under the Typescript namespace
`TypescriptDeclaration`, `TypescriptError`, `TypescriptMember`, `TypescriptType`
and `TypescriptParser` become members of `object Typescript`, taking five names
off the umbrella's exported surface.

The proposal proposed splitting `Declared` out of the declaration's file first.
It cannot be split: `Declared` is `sealed` and the declaration enum is its only
subtype, so the two are pinned to one file — the `SvgDef` rule read in the other
direction. They move together instead, as `Typescript.Declared` and
`Typescript.Declaration`.

`TypescriptDialect` is untouched (reflective lookup). Verified with
`xenophile.test.run` (203 passed).
